requirements.txt
apache-airflow==2.5.1
apache-airflow-providers-google>=8.0.0
apache-airflow-providers-apache-spark>=3.5.0
google-cloud-storage
google-cloud-monitoring
requests
